facet-macros-emit: generate_type_name_fn: move from process_struct
Moved into module `derive`, there may be a better home
(and a better name) for this function.
